+++ models/attribute_values.go
@@ -21,7 +21,9 @@ type AttributeValuesWithDetail struct {
 func (model *AttributeValues) AfterDelete(db *gorm.DB) (err error) {
 	var modelDets = []VariationDetails{}
 	db.Where("attribute_value_id = ?", model.ID).Find(&modelDets)
-	db.Delete(&modelDets)
+	if len(modelDets) > 0 {
+		db.Delete(&modelDets)
+	}
 
 	return
 }
